+++ test/expect/comments_all.js
@@ -12,17 +12,12 @@
 
 //@license: MIT - preserved when the comments option is "all" or "filter"
 
-/*#
-  This comment is preserved when comments option is "all" or "filter"
+/*
+  # Comment preserved when comments option is "all"
+    or "filter" with filter "all" or "titles"
 */
-console.log('with #')     //# keep when comments is "all" or "filter"
+console.log('with #')     //# keep when comments is "all", or "filter" with filter "all" or "titles"
 
-//#ifdef INC
-//#include include1
-//#endif
-
-//#if 1
 console.log('jslint')     //jslint keep this comment with the jslint filter
 console.log('jshint')     // jshint keep this comment with the jshint filter
 console.log('eslint')     // eslint keep with the eslint filter
-//#endif