+++ test/fixtures/define_regex.js
@@ -1,0 +1,4 @@
+//#define $_REGEX /^a'"b"/g
+r=$_REGEX
+//#define $_REGEX new RegExp('^a\'"b"', 'g')
+r=$_REGEX