+++ test/fixtures/defines.js
@@ -1,0 +1,23 @@
+//#define NUM1 1
+//#define NUM2 2
+
+//#ifdef FOO
+'ifdef FOO'
+//#endif
+//#if FOO
+'if FOO'
+//#endif
+//#if defined(FOO)
+'defined FOO'
+//#endif
+
+//#if NUM1 == 1
+'NUM1 == 1'
+//#else
+'no NUM1'
+//#endif
+//#if NUM2
+'NUM2 ok'
+//#else
+'no NUM2'
+//#endif