+++ test/run.js
@@ -1,0 +1,8 @@
+var jasmine = new (require('jasmine'))(),
+    SpecReporter = require('jasmine-spec-reporter'),
+    path = require('path')
+
+jasmine.configureDefaultReporter({print: function () {}})
+jasmine.loadConfigFile(path.join(__dirname, 'jasmine.json'))
+jasmine.addReporter(new SpecReporter())
+jasmine.execute(process.argv.slice(2))